requirements.txt
APScheduler>=3.10.4
feedparser>=6.0.11
python-dotenv>=1.0.1
pydantic>=2.7.4
SQLAlchemy>=2.0.30
requests>=2.32.3
